search: dedicated error for postgres
